chore(react): refactor control components
